mcp-servers/database-inspector/src/requirements.txt
mcp
psycopg2-binary
python-dotenv
pandas